First-start seeds live in bootstrap
One home for what a fresh install needs beyond the schema, called from
both doors (druks init-db and the pytest init_db); seed_harnesses moves
out of user_settings.models with it. Account relationships join inner —
the FK is NOT NULL.
